MLProgress: lower delay to show progressBar
This is linked to commit 89b91b2a in which the delay between updated
from the MediaParsingService was removed. This means that the occasional
null update sent by the medialibrary when changing task will cancel the
task to show the progress bar before it's 1000L delay is done. On low
end devices this delay isn't a problem, but on faster devices the delay
of 1000L causes the progressBar to be canceled before even being shown.